Split token types
