+++ packages/faucet/src/tokens.ts
@@ -1,0 +1,17 @@
+export interface BankTokenMeta {
+  readonly denom: string;
+  /**
+   * The token ticker symbol, e.g. ATOM or ETH.
+   */
+  readonly tickerSymbol: string;
+  /**
+   * The number of fractional digits the token supports.
+   *
+   * A quantity is expressed as atomic units. 10^fractionalDigits of those
+   * atomic units make up 1 token.
+   *
+   * E.g. in Ethereum 10^18 wei are 1 ETH and from the quantity 123000000000000000000
+   * the last 18 digits are the fractional part and the rest the wole part.
+   */
+  readonly fractionalDigits: number;
+}
